[AAOS emulator] Disable Ethernet by default
VirtioWifi exposes network interface as eth0 in Android
guest system and will therefore turn on Ethernet for AAOS
once the usb_host hardware feature is turned on. By design,
Wi-Fi will not be kept as the default network if Ethernet
is present. To workaround the issue, set ethernet_iface_regex
to a value that will never be matched and therefore Ethernet
won't be running.

Bug: 192710732
Signed-off-by: Weilun Du <wdu@google.com>
Change-Id: I34f5ba2160af996801b48e8bf3d658bf00c9c652