Add support for androidboot. configuration of the Audio HAL
By default, the car emulator HAL reads vendor properties, which need
to come from either a device makefile or a script. In some virtualized
environments, it is convenient for the host to adjust the kernel command
line as a source of configuration to the Android guest (same image can
work across hardware and hypervisor, for example)

Allowing Audio HAL to read its configuration from ro.boot.vendor
properties means that these flags can now be configured via
androidboot. properties in the kernel configuration

Bug: 262922694
Test: flash and boot, also tested by partner
Change-Id: I2014e9e8e8e81c78ebcb1cf37a2b64ab080881e2