Fix up property names
In ag/20765521 I added new names for car audio HAL properties, but
I ended up making a typo in the names of two of them.

This CL fixes those names. If we add more properties like these
it may be worthwhile to generate their names programmatically
as to reduce the amount of possible naming errors, but for now
just a manual fix will suffice.

Bug: 262922694
Test: looked at names with care
Change-Id: Ia8045f31667c3467b5205b9d2c1fed081b781482