Show correct display ID in the multi window title.
Due to the different logic to create and manage the display ID in
Android and QEMU, it’s hard to find the correct display when we try to
run an activity on a specific window in the multi display emulator. We
will show the logical display id as the title of multi display window in
qemu. As the logical display id numbering would be based on the sequence
of display in display_layout_configuration.xml, we need to keep the same
sequence with EMULATOR_MULTIDISPLAY_HW_CONFIG and
EMULATOR_MULTIDISPLAY_BOOTANIM_CONFIG value in
device/generic/car/common/car_md.mk.

Bug: 257349071
Test: manual test with the relevant qemu change
Change-Id: I86f022cf623da43cd66a500cb1b930a943345bcc