Update car COD to account for format bits so we can be Car Audio again
Our minor class is 0x8 - Car Audio, but that is supposed to set the 5th
bit to account for the format bits, not the 3rd bit. This is causing our
device to be interpreted as an 0x2 - Handsfree Device by remote devices

This change updates the byte in question, as well as the documentation
so its clearer how COD gets applied by spec, in case we ever need to
change it in the future.

Bug: 312759291
Test: make -j; adb logcat | grep -ie "Using class of device"
Test: Boot device, check HCI snoop logs for HCI COD Write packet
Change-Id: Idbb6b0403b52939663d38f7b4653639a4eec1ba5